requirements.txt
PySide6
pyqtgraph
pyserial==3.5